+++ tests/test_codegen.rs
@@ -104,6 +104,10 @@ pub mod tag {
     include!(\"./tag.rs\");
 }
 
+pub mod test_enum {
+    include!(\"./test_enum.rs\");
+}
+
 pub mod test_nested_array_with_object {
     include!(\"./test_nested_array_with_object.rs\");
 }
@@ -176,7 +180,7 @@ fn test_overridden_path() {
         }
     }
 ",
-        Some(7206),
+        Some(7261),
     );
 }
 
@@ -310,7 +314,8 @@ pub struct OrderListItem {
     pub pet_id: Option<i64>,
     pub quantity: Option<i64>,
 }
-#[derive(Debug, Default, Clone, Serialize, Deserialize)]
+#[derive(Debug, Clone, Serialize, Deserialize)]
+#[allow(non_camel_case_types)]
 pub enum OrderStatus {
     #[serde(rename = \"paymentPending\")]
     PaymentPending,
@@ -321,7 +326,8 @@ pub enum OrderStatus {
     #[serde(rename = \"fulfilled\")]
     Fulfilled,
 }
-#[derive(Debug, Default, Clone, Serialize, Deserialize)]
+#[derive(Debug, Clone, Serialize, Deserialize)]
+#[allow(non_camel_case_types)]
 pub enum OrderTestStringEnum {
     #[serde(rename = \"booya\")]
     Booya,
@@ -330,7 +336,7 @@ pub enum OrderTestStringEnum {
     #[serde(rename = \"true\")]
     True,
     #[serde(rename = \"-53\")]
-    Number_53,
+    Number__53,
 }
 
 impl Order {
@@ -1152,5 +1158,111 @@ impl StatusPutBuilder1<crate::generics::SomeDataFileExists, crate::generics::Foo
 }
 ",
         Some(3154),
+    );
+}
+
+#[test]
+fn test_simple_any_enum() {
+    assert_file_contains_content_at(
+        &(ROOT.clone() + "/tests/test_pet/test_enum.rs"),
+        "#[derive(Debug, Clone)]
+#[allow(non_camel_case_types)]
+pub enum TestEnum {
+    True,
+    Number_1_5,
+    Number_23,
+    Number_964,
+    Number__79_23,
+    Number_14343,
+    Number__964,
+    Hello,
+    Foo,
+    Bar,
+}
+impl serde::Serialize for TestEnum {
+    fn serialize<S: serde::Serializer>(&self, ser: S) -> Result<S::Ok, S::Error> {
+        match self {
+            TestEnum::True => (true).serialize(ser),
+            TestEnum::Number_1_5 => (1.5).serialize(ser),
+            TestEnum::Number_23 => (23).serialize(ser),
+            TestEnum::Number_964 => (964).serialize(ser),
+            TestEnum::Number__79_23 => (-79.23).serialize(ser),
+            TestEnum::Number_14343 => (14343).serialize(ser),
+            TestEnum::Number__964 => (-964).serialize(ser),
+            TestEnum::Hello => (\"hello\").serialize(ser),
+            TestEnum::Foo => (\"foo\").serialize(ser),
+            TestEnum::Bar => (\"bar\").serialize(ser),
+        }
+    }
+}
+impl<'de> serde::Deserialize<'de> for TestEnum {
+    fn deserialize<D: serde::Deserializer<'de>>(deser: D) -> Result<Self, D::Error> {
+        use serde::de::{Error, Unexpected, Visitor};
+        struct VariantVisitor;
+        const EXPECT_MSG: &str = \"valid value for enum TestEnum\";
+
+        impl<'de> Visitor<'de> for VariantVisitor {
+            type Value = TestEnum;
+
+            fn expecting(&self, f: &mut std::fmt::Formatter) -> std::fmt::Result {
+                f.write_str(EXPECT_MSG)
+            }
+
+            fn visit_bool<E: Error>(self, v: bool) -> Result<Self::Value, E> {
+                if v == true {
+                    return Ok(TestEnum::True);
+                }
+                Err(E::invalid_value(Unexpected::Bool(v), &EXPECT_MSG))
+            }
+
+            fn visit_i64<E: Error>(self, v: i64) -> Result<Self::Value, E> {
+                if v == -964 {
+                    return Ok(TestEnum::Number__964);
+                }
+                Err(E::invalid_value(Unexpected::Signed(v), &EXPECT_MSG))
+            }
+
+            fn visit_u64<E: Error>(self, v: u64) -> Result<Self::Value, E> {
+                if v == 23 {
+                    return Ok(TestEnum::Number_23);
+                }
+                if v == 964 {
+                    return Ok(TestEnum::Number_964);
+                }
+                if v == 14343 {
+                    return Ok(TestEnum::Number_14343);
+                }
+                Err(E::invalid_value(Unexpected::Unsigned(v), &EXPECT_MSG))
+            }
+
+            fn visit_f64<E: Error>(self, v: f64) -> Result<Self::Value, E> {
+                if v == 1.5 {
+                    return Ok(TestEnum::Number_1_5);
+                }
+                if v == -79.23 {
+                    return Ok(TestEnum::Number__79_23);
+                }
+                Err(E::invalid_value(Unexpected::Float(v), &EXPECT_MSG))
+            }
+
+            fn visit_str<E: Error>(self, v: &str) -> Result<Self::Value, E> {
+                if v == \"hello\" {
+                    return Ok(TestEnum::Hello);
+                }
+                if v == \"foo\" {
+                    return Ok(TestEnum::Foo);
+                }
+                if v == \"bar\" {
+                    return Ok(TestEnum::Bar);
+                }
+                Err(E::invalid_value(Unexpected::Str(v), &EXPECT_MSG))
+            }
+        }
+
+        deser.deserialize_any(VariantVisitor)
+    }
+}
+",
+        Some(0),
     );
 }